passwdqc_check: is_based(): Always discount non-words from passphrases
This was previously limited to our hard-coded character sequences, but is
now extended to any non-purely-alphabetic original substrings.

With this and our current default policy, the number of accepted common
passwords from JtR default password.lst reduces from 1657 to 1640 (out of
1795690), while the number of accepted CMIYC-2010 uncracked passwords
stays unchanged at 6333 (out of 16324).  The real effect is with external
wordlists containing many alternating character class strings, which we
previously skipped substring-matching passphrase-alikes against because of
the limitation.